pyproject.toml
[tool.poetry]
name = "unity-realtime-npc"
version = "0.1.0"
description = "A python-based client for OpenAI's Realtime API for Unity"
authors = ["Nestor Almeida <nestor.almeida@proyectran.com>"]
license = "MIT"
readme = "app/README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.13,<3.14"
llama-index-core = "0.14.2"
websockets = "15.0.1"
python-dotenv = "1.1.1"
tzdata = ">=2025.2"
fastapi = "0.116.1"
uvicorn = "0.35.0"
openai = "1.107.1"
openai-agents = "0.3.0"
llama-index-vector-stores-qdrant = "0.8.5"
llama-index-llms-openai = "0.5.6"
llama-index-embeddings-openai = "0.5.1"
llama-index-agent-openai = "0.4.12"
fastembed-gpu = { version = "0.7.3", optional = true }
transformers = { version = "4.56.1", extras = ["torch"], optional = true }
arize-phoenix = "11.34.0"
openinference-instrumentation-llama_index = "4.3.5"
llama-index-callbacks-openinference = "0.4.1"

[tool.poetry.extras]
fastembed-gpu = ["fastembed-gpu"]
transformers = ["transformers"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
